requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
pytz==2023.3
cryptography==41.0.7

